Refactor header/footer API to remove indexPath
